transfer color/element.tsx into index.tsx
